+++ NEWS.md
@@ -10,6 +10,7 @@ Major internal refactor with many small feature improvements:
 - New tests for batch functions
 - Bug fixed where the system prompts did not work in `claude()` batch requests
 - New `field_object()` for tidyllm schemata
+- Batch API implemented for `groq()`
 
 # Dev-Version 0.3.3